Switch to Universal-ctags and delta diffs
Fancy diffs are nice, but a maintained ctags is even better!

NOTE: rbenv-ctags still has a dependency on the ctags (which is really
the abandoned Exuberant ctag project), which will be unmet. But that's
OK b/c it will find `ctags` (universal-ctags) on the $PATH and use it.
So Brew Doctor will complain, but everything will still work just fine.