Default to full path to rbenv/nodenv_path
When doing a Homebrew install of these tools, Homebrew is nice enough to
add their binaries to our path (b/c they live on the Homebrew binary
path). But when installing them via git (e.g., rbenv-installer, on a
Linux dev box), we need to use the full path to the binaries to do the
shell initialization.