Use whatever fzf is available on $PATH
Prior to this we were installing fzf via Homebrew (on macOS) or apt
(Linux) AND also letting the fzf Vim Plug install another version. But
then we'd adjust our $PATH to always use the one being installed by Vim.
Now we just expect one to be installed already (Homebrew or Apt) and use
it.