Remove /usr/local/bin munging
This is no longer needed on Apple Silicon since Homebrew puts binaries
in /opt/homebrew/bin, not /usr/local/bin. Secondly, having this in front
of /opt/homebrew/bin can cause issues when older binaries/libs are
there, but we intend to be using newer versions installed by Homebrew.
So, let's 🔥 this down.